refactor: moving the monitor clients
